requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0          # Load .env variables

# ── Data Processing ───────────────────────────────────────────────────────────
polars>=0.20.0                # Fast DataFrame library (replaces pandas for performance)
pandas>=2.1.0                 # Still needed by some ML/finance libs
numpy>=1.26.0
scipy>=1.12.0

# ── Market Data ───────────────────────────────────────────────────────────────
alpaca-py>=0.20.0             # Alpaca Markets SDK (data + trading)
yfinance>=0.2.36              # Yahoo Finance fallback for daily data
requests>=2.31.0
pyarrow>=15.0.0               # Parquet read/write support for data caching

# ── Machine Learning ──────────────────────────────────────────────────────────
# torch>=2.2.0 (CPU-only — DO NOT use plain pip install)
# Install CUDA build instead:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
torch>=2.2.0                  # PyTorch - install via cu128 index above for GPU support
scikit-learn>=1.4.0           # Preprocessing, metrics, baseline models
xgboost>=2.0.0                # Gradient boosting baseline

# ── Backtesting ───────────────────────────────────────────────────────────────
vectorbt>=0.26.0              # Vectorized backtesting (fast, pandas-based)

# ── Feature Engineering ───────────────────────────────────────────────────────
ta>=0.11.0                    # Technical analysis indicators (RSI, MACD, BB, etc.)

# ── Visualization ─────────────────────────────────────────────────────────────
plotly>=5.18.0
matplotlib>=3.8.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml>=6.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
